+++ models/entities/Users.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace app\models;
+
+
+use app\models\entities\DataEntity;
+
+class Users extends DataEntity
+{
+    public $id;
+    public $login;
+    public $pass;
+
+}
