+++ models/repositories/UsersRepository.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Created by IntelliJ IDEA.
+ * User: iluka
+ * Date: 04.06.2019
+ * Time: 15:45
+ */
+
+namespace app\models\repositories;
+
+
+class UsersRepository
+{
+
+    public static function getTableName()
+    {
+        return 'users';
+    }
+}
